Create field validation
Generate 40 points randomly across 4 levels of habitat thresholds. Also identify which IDNR parks will be used.
Note: set.seed() was used to generate the random points so it can be replicated. Erase function in arcmap was used to erase 3 points that were within 1km of presence locations to prevent pseudo-replication. All 3 were high suitability with 2 in Mason county and 1 in Cass County